requirements.txt
boto3
python-dotenv
AWSIoTPythonSDK
Flask
Flask-Cors
streamlit
plotly
